+++ moshi-ir/moshi-compiler-plugin/src/main/kotlin/dev/zacsweers/moshix/ir/compiler/api/TargetProperty.kt
@@ -5,6 +5,7 @@ package dev.zacsweers.moshix.ir.compiler.api
 import dev.zacsweers.moshix.ir.compiler.util.type
 import org.jetbrains.kotlin.ir.declarations.IrProperty
 import org.jetbrains.kotlin.ir.types.IrType
+import org.jetbrains.kotlin.ir.util.TypeRemapper
 
 /** A property in user code that maps to JSON. */
 internal data class TargetProperty(
@@ -12,18 +13,21 @@ internal data class TargetProperty(
   val parameter: TargetParameter?,
   val jsonName: String?,
   val jsonIgnore: Boolean,
+  val type: IrType = parameter?.type ?: property.type,
 ) {
   val name: String
     get() = property.name.identifier
-
-  val type: IrType
-    get() = parameter?.type ?: property.type
 
   val parameterIndex: Int
     get() = parameter?.index ?: -1
 
   val hasDefault: Boolean
     get() = parameter?.hasDefault ?: true
+
+  internal fun remapTypes(remapper: TypeRemapper): TargetProperty {
+    val remappedParameter = parameter?.remapTypes(remapper)
+    return copy(parameter = remappedParameter, type = remapper.remapType(type))
+  }
 
   override fun toString(): String = name
 }